+++ spec/lib/slkecho/blocks_builder_spec.rb
@@ -9,10 +9,10 @@ RSpec.describe Slkecho::BlocksBuilder do
       let(:blocks) do
         [
           {
-            "type" => "section",
-            "text" => {
-              "type" => "mrkdwn",
-              "text" => "message"
+            type: "section",
+            text: {
+              type: "mrkdwn",
+              text: "message"
             }
           }
         ]