+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - refactor: Parse JSON with symbolize names
 - refactor: Make request body with symbolize keys
 - refactor: Make blocks with symbolize keys
+- refactor: Extract http response checking
 
 ## [v2.1.0] - 2024-01-19