+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - refactor: Parse JSON with symbolize names
 - refactor: Make request body with symbolize keys
 - refactor: Make blocks with symbolize keys
+- refactor: Extract http response checking
+- test: Introduce saharspec gem
 
 ## [v2.1.0] - 2024-01-19